requirements.txt
fastapi
uvicorn
uvloop>=0.14.0
httptools==0.1.*
pydantic
uvicorn[standard]
requests
streamlit
pillow
pandas
mysql
mysql.connector
